Bug 1727453 - Expose merge-analyses for testing purposes.
This extracts  most of the logic in `bin/merge-analyses.rs` into
`file_formater/merger.rs`, leaving most of the logic intact,
specifically:
- It continues to operate on absolute paths.
- It continues to consume the files itself, including all JSON parsing.
- It continues to output to std::io::Write, but now it's not implicitly
  to stdout and instead is provided by the caller.

A new merge-analyses searchfox-tool command is then introduced, but
it's a relatively thin wrapper.  It transforms tree-relative analysis
paths into absolute paths, consumes the resulting string output and
parses it back into the standard pipeline JsonRecords/JsonRecordsByFile
rep.
