+++ tools/src/cmd_pipeline/mod.rs
@@ -6,6 +6,7 @@ pub mod interface;
 pub mod parser;
 
 mod cmd_filter_analysis;
+mod cmd_merge_analyses;
 mod cmd_prod_filter;
 mod cmd_query;
 mod cmd_show_html;
