+++ tools/src/file_format/mod.rs
@@ -1,2 +1,3 @@
 pub mod analysis;
 pub mod identifiers;
+pub mod merger;
